+++ app/src/main/java/com/artemklymenko/sneakersapp/pages/product/ProductUiEvent.kt
@@ -3,7 +3,7 @@ package com.artemklymenko.sneakersapp.pages.product
 import com.artemklymenko.sneakersapp.core.base.common.events.UiEvent
 
 sealed class ProductUiEvent : UiEvent {
-    data class LoadData(val id: Long) : ProductUiEvent()
+    data class LoadData(val id: Int) : ProductUiEvent()
     data object AddToCartOnClick : ProductUiEvent()
     data object UpdateFavouriteOnClick: ProductUiEvent()
 }
